refactor: separate parkMark
